requirements.txt
# Core dependencies
google-genai>=1.0.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0

# Embedding and NLP (LOCAL - no API calls)
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# BM25 for sparse retrieval
rank-bm25>=0.2.2

# Date/Time handling
python-dateutil>=2.8.0

# Utilities
pydantic>=2.0.0
rich>=13.0.0
